backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
httpx>=0.26.0
supabase>=2.3.0
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
websockets>=12.0
speechmatics-python>=1.0.0


